updating production: mailing list group

+++ mod_app/utils/mixins.py
@@ -1,3 +1,4 @@
+from django.conf import settings
 from django.template.loader import get_template
 from django.utils.html import format_html
 
@@ -50,21 +51,26 @@ class s3BrowserButtonMixin:
         self.readonly_fields += ("s3_browser_button",)
 
     def s3_browser_button(self, obj):
-        button_html = get_template("components/s3_browse_button.html")
-        return button_html.render()
+        if settings.ENVIRONMENT != "local":
+            button_html = get_template("components/s3_browse_button.html")
+            return button_html.render()
+        else:
+            button_html = "Not available locally"
+            return button_html
 
 
 class EmailMixin:
     def save_model(self, request, obj, form, change):
         super().save_model(request, obj, form, change)
-
-        # Check if it's a new instance or an update
-        if change:
-            build_and_send_email(request, obj, "updated")
-        else:
-            # new instance
-            build_and_send_email(request, obj, "added")
+        if settings.ENVIRONMENT != "local":
+            # Check if it's a new instance or an update
+            if change:
+                build_and_send_email(request, obj, "updated")
+            else:
+                # new instance
+                build_and_send_email(request, obj, "added")
 
     def delete_model(self, request, obj):
-        build_and_send_email(request, obj, "deleted")
-        super().delete_model(request, obj)
+        if settings.ENVIRONMENT != "local":
+            build_and_send_email(request, obj, "deleted")
+            super().delete_model(request, obj)
